# api/openapi.yml
openapi: 3.0.3
info:
  title: Robokache
  description: Large object document store for ROBOKOP
  version: 3.0.0
  contact:
    email: patrick@covar.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
security:
  - google: []
paths:
  /api/document:
    get:
      summary: Get documents
      description: Get a list of documents that are either public or associated with the current user
      responses:
        '200':
          description: Documents
          content:
            application/json:
              schema:
                type: array
                # Include ID along with the document
                items:
                  allOf:
                   - $ref: '#/components/schemas/Document'
                   - properties:
                       id:
                         type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      summary: Create a document
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '201':
          $ref: '#/components/responses/IdOfCreated'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/document/{id}:
    get:
      summary: Get document by ID
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NoSuchDocument'
    put:
      summary: Update fields of document
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '200':
          description: Successfully updated
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NoSuchDocument'
    delete:
      summary: Delete document by ID
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NoSuchDocument'
  /api/document/{id}/children:
    get:
      summary: Get documents that have this document as a parent
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Document'
                    - properties:
                        id:
                          type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      summary: Create a new child document and set data field
      description: Shorthand to create a child document and set the data field using only one route. This creates the document with default values for fields.
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        description: Data object
        content:
          application/json:
            schema:
              type: string
      responses:
        '200':
          $ref: '#/components/responses/IdOfCreated'
  /api/document/{id}/data:
    get:
      summary: Get the data associated with this document
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Data object
          content:
            application/json:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      summary: Set the data associated with this document
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        description: Data object
        content:
          application/json:
            schema:
              type: string
      responses:
        '200':
          description: Successfully Updated
        '401':
          $ref: '#/components/responses/UnauthorizedError'

components:
  schemas:
    Document:
      properties:
        parent:
          type: string
        visibility:
          type: integer
  securitySchemes:
    google:
      type: http
      scheme: bearer
      bearerFormat: jwt
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
    NoSuchDocument:
      description: Document ID not found
    IdOfCreated:
      description: ID of created document
      content:
        application/text:
          example: 1234
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string